serial: msm_geni_serial: Fix console log false start issue
With the commit d0e6e429d99a after merging android-4.14-stable
(4.14.257), it is possible to re-configure the preferred console
when the serial ports are added even if the cmdline parameter
'console' is set as null.

Set the flag 'console_set_on_cmdline' as 1 directly since
we do not need to setup the console at runtime.

Bug: 211946493
Bug: 210364486
Change-Id: I2f472520ad5ff8593878cfd4f818febc965d828e
Signed-off-by: JohnnLee <johnnlee@google.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: firebird11 <hbgassel@gmail.com>